fix(editor): leave a remote peer's link space to that peer
Review catch: the changed-range guard keeps an edit *elsewhere* from
rewriting marks at the cursor, but a remote peer typing the space at that
same spot still triggered a local exit — pushing our normalisation into
their typing, on top of the exit their own editor already replicated.

`list-join` already had this guard for the same reason (two peers rewriting
one range converges on the wrong shape). Extract its `isRemoteChange` into
`shared/collaboration.ts` rather than duplicating twenty lines of
y-prosemirror plugin-key sniffing, and use it in both.

Moving the helper also types `PluginKey`'s runtime `key` field explicitly,
which drops a pre-existing type-check error in `list-join`.
